Migrate to new API
Also:
  * improve method names
